Compile gen_struct_info in a hermetic environment
At least prevent `EMMAKEN_CFLAGS` from influencing it.
If `gen_struct_info` is run with `EMMAKEN_CFLAGS=-O2` it
will fail with undefined symbol `iprintf`.

+++ tools/gen_struct_info.py
@@ -275,7 +275,7 @@ def inspect_headers(headers, cflags):
 
   show(shared.shlex_join(cmd))
   try:
-    subprocess.check_call(cmd)
+    subprocess.check_call(cmd, env=system_libs.clean_env())
   except subprocess.CalledProcessError as e:
     sys.stderr.write('FAIL: Compilation failed!: %s\n' % e.cmd)
     sys.exit(1)
@@ -284,11 +284,12 @@ def inspect_headers(headers, cflags):
   show('Calling generated program... ' + js_file[1])
   info = shared.run_js_tool(js_file[1], stdout=shared.PIPE).splitlines()
 
-  # Remove all temporary files.
-  os.unlink(src_file[1])
+  if not DEBUG:
+    # Remove all temporary files.
+    os.unlink(src_file[1])
 
-  if os.path.exists(js_file[1]):
-    os.unlink(js_file[1])
+    if os.path.exists(js_file[1]):
+      os.unlink(js_file[1])
 
   # Parse the output of the program into a dict.
   return parse_c_output(info)
